Split load generator into modules
